move EventSourcing.ForgettablePayloads.Persistence.PostgreSql DI code into main EventSourcing.ForgettablePayloads.Persistence.PostgreSql package
