--- EventSourcing.ForgettablePayloads/Persistence/PostgreSql/Extensions.DependencyInjection.PostgreSql/Extensions.DependencyInjection.PostgreSql.csproj
@@ -1,37 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <TreatWarningsAsErrors Condition=" '$(Configuration)' == 'Release' ">true</TreatWarningsAsErrors>
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-        <RootNamespace>EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql</RootNamespace>
-        <AssemblyName>EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql</AssemblyName>
-    </PropertyGroup>
-
-    <!-- NuGet metadata -->
-    <PropertyGroup>
-        <Title>EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql</Title>
-        <Description>This package contains DependencyInjection extensions for PostgreSql implementation of EventSourcing.ForgettablePayloads.Persistence package.</Description>
-        <Authors>Dawid Szmidka</Authors>
-        <PackageId>EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql</PackageId>
-        <Version>0.1.0-alpha</Version>
-        <PackageReleaseNotes>
-            - Initial version of EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql project
-        </PackageReleaseNotes>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <PackageTags>EventSourcing;EventSourcing.ForgettablePayloads.Extensions;EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql</PackageTags>
-        <IncludeSymbols>true</IncludeSymbols>
-        <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-        <IncludeSource>true</IncludeSource>
-        <RepositoryUrl>https://github.com/TheName/EventSourcing</RepositoryUrl>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="5.0.0" />
-    </ItemGroup>
-    
-    <ItemGroup>
-      <ProjectReference Include="..\ForgettablePayloads.Persistence.PostgreSql\ForgettablePayloads.Persistence.PostgreSql.csproj" />
-    </ItemGroup>
-
-</Project>
