+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/ForgettablePayloads.Persistence.PostgreSql.IntegrationTests/PostgreSqlCollectionFixture.cs
@@ -7,8 +7,8 @@ using EventSourcing.Extensions;
 using EventSourcing.Extensions.DependencyInjection.Serialization.Json;
 using EventSourcing.ForgettablePayloads.Extensions;
 using EventSourcing.ForgettablePayloads.Extensions.DatabaseMigrations.Persistence.PostgreSql.DbUp.Extensions;
-using EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql;
 using EventSourcing.ForgettablePayloads.Persistence.PostgreSql;
+using EventSourcing.ForgettablePayloads.Persistence.PostgreSql.Extensions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
