+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/ForgettablePayloads.Persistence.PostgreSql/Extensions/EventSourcingForgettablePayloadsPersistenceBuilderExtensions.cs
@@ -1,12 +1,10 @@
 ﻿using System;
-using EventSourcing.ForgettablePayloads.Persistence;
 using EventSourcing.ForgettablePayloads.Persistence.Abstractions.DependencyInjection;
-using EventSourcing.ForgettablePayloads.Persistence.PostgreSql;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
 
-namespace EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql
+namespace EventSourcing.ForgettablePayloads.Persistence.PostgreSql.Extensions
 {
     /// <summary>
     /// The <see cref="IEventSourcingForgettablePayloadsPersistenceBuilder"/> extensions.
